refactor: extract Calendar
